proso_user: support for registration
The main goal of this commit is to merge proso_auth and proso_user. I
have tried to implement the most of functionality of proso_auth on the
backend side, but there is also a part which has to be added to
proso_models. The refactoring does not contain the source code for
client side, so I have not deleted the proso_auth yet.